+++ src/FormidableError.js
@@ -7,12 +7,14 @@ const filenameNotString = 1005;
 const maxFieldsSizeExceeded = 1006;
 const maxFieldsExceeded = 1007;
 const smallerThanMinFileSize = 1008;
-const biggerThanMaxFileSize = 1009;
+const biggerThanTotalMaxFileSize = 1009;
 const noEmptyFiles = 1010;
 const missingContentType = 1011;
 const malformedMultipart = 1012;
 const missingMultipartBoundary = 1013;
 const unknownTransferEncoding = 1014;
+const maxFilesExceeded = 1015;
+const biggerThanMaxFileSize = 1016;
 const pluginFailed = 1017;
 
 const FormidableError = class extends Error {
@@ -32,6 +34,7 @@ export {
   filenameNotString,
   maxFieldsSizeExceeded,
   maxFieldsExceeded,
+  maxFilesExceeded,
   smallerThanMinFileSize,
   biggerThanMaxFileSize,
   noEmptyFiles,
@@ -39,6 +42,7 @@ export {
   malformedMultipart,
   missingMultipartBoundary,
   unknownTransferEncoding,
+  biggerThanTotalMaxFileSize,
   pluginFailed,
 };